- reorganized values/types (removing types) - added converters for all types, replacing the utility methods previously used - created a converter registry - created a erd serializer, adjusted the appliance class to leverage
