+++ gekitchen/clients/base_client.py
@@ -6,8 +6,7 @@ import logging
 from aiohttp import ClientSession
 from typing import Any, Callable, Dict, Optional, Tuple
 from ..const import EVENT_APPLIANCE_INITIAL_UPDATE
-from ..erd.erd_codes import ErdCode
-from ..erd_utils import ErdCodeType
+from ..erd import ErdCode, ErdCodeType
 from ..exc import GeNotAuthedError
 from ..ge_appliance import GeAppliance
