+++ gekitchen/clients/websocket_client.py
@@ -19,8 +19,7 @@ from ..const import (
     EVENT_DISCONNECTED,
     EVENT_GOT_APPLIANCE_LIST,
 )
-from ..erd import ErdCode
-from ..erd_utils import ErdCodeType
+from ..erd import ErdCode, ErdCodeType
 from ..ge_appliance import GeAppliance
 from ..login_flow import do_full_wss_flow
