--- gekitchen/erd/types/__init__.py
@@ -1,4 +1,0 @@
-"""GE ERD Types"""
-
-from .fridge import *
-from .oven import *
